% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{install_tobac}
\alias{install_tobac}
\title{Install Tobac}
\usage{
install_tobac(envname = "harp-py", new_env = identical(envname, "harp-py"))
}
\arguments{
\item{envname}{The virtual environment into which to install the tobac Python
modules. Default is "harp-py".}

\item{new_env}{Whether to start from a new environment by removing the
\code{envname} if it already exists. The defulat is TRUE if \code{envname} already
exists, and you will be prompted if you want to remove the virtualenv
\code{envname}.}
}
\description{
\code{install_tobac()} installs the Python modules needed to run Tobac functions.
The packages are installed in a \emph{virtualenv}, which by default is named
"harp-py". This virtual environment is the default virtual environment in
which all harp related Python modules are installed. If you already have the
"harp-py" virtualenv, you will be asked if you want to remove the virtualenv
prior to installation. Typically you would answer no to this.
}
