% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindings.R
\name{link_tracks}
\alias{link_tracks}
\title{harp wrapper to the tobac function \code{tracking.linking_trackpy()}.}
\usage{
link_tracks(
  features,
  field_data,
  dz = NULL,
  data_col = "gridded_data",
  dttm_col = "valid_dttm",
  d_max = NULL,
  subnetwork_size = NULL,
  v_max = NULL,
  memory = 0,
  stubs = 1,
  time_cell_min = NULL,
  order = 1,
  extrapolate = 0,
  method_linking = c("predict", "random"),
  adaptive_step = NULL,
  adaptive_stop = NULL,
  cell_number_start = 1,
  cell_number_unassigned = -1,
  vertical_coord = NULL,
  min_h1 = NULL,
  max_h1 = NULL,
  min_h2 = NULL,
  max_h2 = NULL,
  pbc_flag = c("none", "hdim_1", "hdim_2", "both")
)
}
\arguments{
\item{features}{A features data frame as returned by
\code{\link{detect_features_multithreshold()}} or
\code{\link{segment_2d()}}}

\item{field_data}{A \code{harp_grid_df} data frame such as one returned by the
\emph{harpIO} functions
\code{\link[harpIO:read_grid]{harpIO::read_grid()}} with \code{data_frame = TRUE}, \code{\link[harpIO:read_forecast]{harpIO::read_forecast()}},
or \code{\link[harpIO:read_analysis]{harpIO::read_analysis()}}.}

\item{dz}{Constant vertical grid spacing (meters), optional. If not specified
and the input is 3D, this function requires that vertical_coord is
available in the features input. If you specify a value here, this function
assumes that it is the constant z spacing between points, even if
\code{vertical_coord} is specified.}

\item{data_col}{<\code{\link[=dplyr_tidy_select]{tidy-select}}> The column in
\code{field_data} containing the fields to be used to detect features in. Should
be a \verb{<geolist>} column. If the named column is not found in \code{field_data},
but \code{field_data} contains 1 \verb{<geolist>} column, that \verb{<geolist>} column is
used.}

\item{dttm_col}{<\code{\link[=dplyr_tidy_select]{tidy-select}}> The column in
\code{field_data} containing the date-times to be used for the time dimension.
Can be numeric with units in Unix time (seconds since 1970-01-01 00:00:00),
or a \verb{<POSIXt>}column. If the named column is not found in \code{field_data},
but \code{field_data} contains 1 \verb{<POSIXt>} column, that \verb{<POSIXt>} column is
used.}

\item{d_max}{Maximum search range in metres. Only one of d_max, or v_max can
be set. Default is \code{NULL}.}

\item{subnetwork_size}{Maximum size of subnetwork for linking. This parameter
should be adjusted when using adaptive search. Usually a lower value is
desired in that case. For a more in depth explanation have look
\href{https://soft-matter.github.io/trackpy/v0.5.0/tutorial/adaptive-search.html}{here}.
If \code{NULL}, 30 is used for regular search and 15 for adaptive search.
Default is \code{NULL}.}

\item{v_max}{Speed at which features are allowed to move in m/s. Only one of
d_max or v_max can be set. Default is \code{NULL}.}

\item{memory}{Number of output timesteps features allowed to vanish for to
still be considered tracked. Default is 0. \emph{Warning} :: This parameter
should be used with caution, as it can lead to erroneous trajectory
linking, espacially for data with low time resolution.}

\item{stubs}{Minimum number of timesteps of a tracked cell to be reported.
Default is 1}

\item{time_cell_min}{Minimum length in time that a cell must be tracked for
to be considered a valid cell in seconds. Default is \code{NULL}.}

\item{order}{Order of polynomial used to extrapolate trajectory into gaps and
on start and end point. Default is 1.}

\item{extrapolate}{Number or timesteps to extrapolate trajectories.
\emph{Currently unused}. Default is 0.}

\item{method_linking}{Flag choosing method used for trajectory linking. Can
be "predict" or "random". Default is ‘predict’. \emph{Note}: the default in
tobac is "random" though they typically encourage users to use ‘predict’.}

\item{adaptive_step}{Reduce search range by multiplying it by this factor.
Needs to be used in combination with \code{adaptive_stop}. Default is \code{NULL}.}

\item{adaptive_stop}{If not \verb{NULL, when encountering an oversize subnet, retry by progressively reducing }search_range\verb{by multiplying with}adaptive_step\verb{until the subnet is solvable. If search_range becomes <=}adaptive_stop\verb{, give up and raise a SubnetOversizeException. Needs to be used in combination with }adaptive_step\verb{. Default is }NULL`.}

\item{cell_number_start}{Cell number for first tracked cell. Default is 1.}

\item{cell_number_unassigned}{Number to set the unassigned/non-tracked cells
to. Default is -1.}

\item{vertical_coord}{Name of the vertical coordinate. The vertical
coordinate used must be metres. If \verb{NULL, it tries to auto-detect. It looks for the coordinate or the dimension name corresponding to the string. To use }dz\verb{, set this to }NULL`.}

\item{min_h1}{Minimum hdim_1 value, required when \code{pbc_flag} is "hdim_1" or
"both".}

\item{max_h1}{Maximum hdim_1 value, required when \code{pbc_flag} is "hdim_1" or
"both".}

\item{min_h2}{Minimum hdim_2 value, required when \code{pbc_flag} is "hdim_2" or
"both".}

\item{max_h2}{Maximum hdim_2 value, required when \code{pbc_flag} is "hdim_2" or
"both".}

\item{pbc_flag}{Sets whether to use periodic boundaries, and if so in which
directions. "none" means that we do not have periodic boundaries "hdim_1"
means that we are periodic along hdim1, "hdim_2" means that we are periodic
along hdim2 and "both" means that we are periodic along both horizontal
dimensions}
}
\value{
Data frame of the linked features, containing the column "cell", with
integers indicating the affiliation of a feature to a specific track, and
the variable "time_cell" with the time, in seconds' the cell has already
existed.
}
\description{
harp wrapper to the tobac function \code{tracking.linking_trackpy()}.
}
