% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindings.R
\name{segment_2d}
\alias{segment_2d}
\title{harp wrapper for tobac \code{segmentation.segmentation_2d()} function}
\usage{
segment_2d(
  features,
  field_data,
  threshold,
  data_col = "gridded_data",
  dttm_col = "valid_dttm",
  target = c("max", "min"),
  level = NULL,
  method = "watershed",
  max_distance = NULL,
  pbc_flag = c("none", "h_dim1", "h_dim2", "both"),
  seed_3d_flag = c("column", "box"),
  statistic = NULL
)
}
\arguments{
\item{features}{A features data frame, as returned by
\code{\link{detect_features_multithreshold()}}.}

\item{field_data}{A \code{harp_grid_df} data frame such as one returned by the
\emph{harpIO} functions
\code{\link[harpIO:read_grid]{harpIO::read_grid()}} with \code{data_frame = TRUE}, \code{\link[harpIO:read_forecast]{harpIO::read_forecast()}},
or \code{\link[harpIO:read_analysis]{harpIO::read_analysis()}}.}

\item{threshold}{Threshold for the watershedding field to be used for the
mask.}

\item{data_col}{<\code{\link[=dplyr_tidy_select]{tidy-select}}> The column in
\code{field_data} containing the fields to be used to detect features in. Should
be a \verb{<geolist>} column. If the named column is not found in \code{field_data},
but \code{field_data} contains 1 \verb{<geolist>} column, that \verb{<geolist>} column is
used.}

\item{dttm_col}{<\code{\link[=dplyr_tidy_select]{tidy-select}}> The column in
\code{field_data} containing the date-times to be used for the time dimension.
Can be numeric with units in Unix time (seconds since 1970-01-01 00:00:00),
or a \verb{<POSIXt>}column. If the named column is not found in \code{field_data},
but \code{field_data} contains 1 \verb{<POSIXt>} column, that \verb{<POSIXt>} column is
used.}

\item{target}{Flag to determine if tracking is targeting minima or maxima in
the data. Should be "max" or "min". Default is "max".}

\item{level}{Levels at which to seed the cells for the watershedding
algorithm. Default is \code{NULL}.}

\item{method}{Flag determining the algorithm to use (currently only
"watershed" is implemented).}

\item{max_distance}{Maximum distance from a marker allowed to be classified
as belonging to that cell in metres. Default is \code{NULL}.}

\item{pbc_flag}{Sets whether to use periodic boundaries, and if so in which
directions. "none" means that we do not have periodic boundaries "hdim_1"
means that we are periodic along hdim1, "hdim_2" means that we are periodic
along hdim2 and "both" means that we are periodic along both horizontal
dimensions}

\item{seed_3d_flag}{Seed 3D field at feature positions with either the full
column (default) or a box of user-set size.}

\item{statistic}{Default is \code{NULL}. Parameter to calculate bulk statistics
within feature detection. Named list with callable function(s) to apply
over the region of each detected feature and the name of the statistics to
appear in the feature output data frame. The functions should be the values
and the names of the metric the keys (e.g. list(mean = np.mean)). This
probably doesn't work yet as there needs to be a translation between R and
Python functions here.}
}
\value{
A named list with:
\itemize{
\item segments: A data frame with a \code{geolist} column with 2d fields on the
same domain as the input data with integer values denoting each segment.
\item features: Feature data frame including the number of cells in the
segmented area of the feature at the time step.
}
}
\description{
harp wrapper for tobac \code{segmentation.segmentation_2d()} function
}
