% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{toabc}
\alias{toabc}
\title{Main tobac module}
\format{
Tobac module
}
\usage{
toabc
}
\description{
Interface to main Tobac module. Provides access to top level functions as
well as sub-modules using \code{$} instead of \code{.} as in Python.
}
\examples{
\dontrun{
library(tobac)

tobac$sys$api_version
}

}
\keyword{datasets}
